requirements.txt
robotframework
robotframework-requests
robotframework-databaselibrary
robotframework-seleniumlibrary

PyYAML==6.0